Fix more feature flags and refactor tests
