+++ crates/icrate/src/Foundation/additions/attributed_string.rs
@@ -4,10 +4,7 @@ use objc2::rc::{DefaultId, Id, Shared};
 use objc2::runtime::Object;
 use objc2::{extern_methods, ClassType};
 
-use crate::Foundation::{
-    NSAttributedString, NSAttributedStringKey, NSCopying, NSDictionary, NSMutableAttributedString,
-    NSMutableCopying, NSString,
-};
+use crate::Foundation::{self, NSAttributedString, NSAttributedStringKey};
 
 // SAFETY: `NSAttributedString` is immutable and `NSMutableAttributedString`
 // can only be mutated from `&mut` methods.
@@ -34,16 +31,19 @@ extern_methods!(
         ///
         /// The attributes must be valid.
         #[doc(alias = "initWithString:")]
+        #[cfg(feature = "Foundation_NSDictionary")]
+        #[cfg(feature = "Foundation_NSString")]
         pub unsafe fn new_with_attributes(
-            string: &NSString,
-            attributes: &NSDictionary<NSAttributedStringKey, Object>,
+            string: &Foundation::NSString,
+            attributes: &Foundation::NSDictionary<NSAttributedStringKey, Object>,
         ) -> Id<Self, Shared> {
             unsafe { Self::initWithString_attributes(Self::alloc(), string, Some(attributes)) }
         }
 
         /// Creates a new attributed string without any attributes.
         #[doc(alias = "initWithString:")]
-        pub fn from_nsstring(string: &NSString) -> Id<Self, Shared> {
+        #[cfg(feature = "Foundation_NSString")]
+        pub fn from_nsstring(string: &Foundation::NSString) -> Id<Self, Shared> {
             Self::initWithString(Self::alloc(), string)
         }
     }
@@ -55,94 +55,5 @@ impl DefaultId for NSAttributedString {
     #[inline]
     fn default_id() -> Id<Self, Self::Ownership> {
         Self::new()
-    }
-}
-
-unsafe impl NSCopying for NSAttributedString {
-    type Ownership = Shared;
-    type Output = NSAttributedString;
-}
-
-unsafe impl NSMutableCopying for NSAttributedString {
-    type Output = NSMutableAttributedString;
-}
-
-impl alloc::borrow::ToOwned for NSAttributedString {
-    type Owned = Id<NSAttributedString, Shared>;
-    fn to_owned(&self) -> Self::Owned {
-        self.copy()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use alloc::string::ToString;
-    use alloc::{format, vec};
-
-    use objc2::rc::{autoreleasepool, Owned};
-
-    use super::*;
-    use crate::Foundation::NSObject;
-
-    #[test]
-    fn test_new() {
-        let s = NSAttributedString::new();
-        assert_eq!(&s.string().to_string(), "");
-    }
-
-    #[test]
-    fn test_string_bound_to_attributed() {
-        let attr_s = {
-            let source = NSString::from_str("Hello world!");
-            NSAttributedString::from_nsstring(&source)
-        };
-        let s = autoreleasepool(|_| attr_s.string());
-        assert_eq!(s.len(), 12);
-    }
-
-    #[test]
-    fn test_from_nsstring() {
-        let s = NSAttributedString::from_nsstring(&NSString::from_str("abc"));
-        assert_eq!(&s.string().to_string(), "abc");
-    }
-
-    #[test]
-    fn test_copy() {
-        let s1 = NSAttributedString::from_nsstring(&NSString::from_str("abc"));
-        let s2 = s1.copy();
-        // NSAttributedString performs this optimization in GNUStep's runtime,
-        // but not in Apple's; so we don't test for it!
-        // assert_eq!(Id::as_ptr(&s1), Id::as_ptr(&s2));
-        assert!(s2.is_kind_of::<NSAttributedString>());
-
-        let s3 = s1.mutable_copy();
-        assert_ne!(Id::as_ptr(&s1), Id::as_ptr(&s3).cast());
-        assert!(s3.is_kind_of::<NSMutableAttributedString>());
-    }
-
-    #[test]
-    fn test_debug() {
-        let s = NSAttributedString::from_nsstring(&NSString::from_str("abc"));
-        let expected = if cfg!(feature = "gnustep-1-7") {
-            "abc{}"
-        } else {
-            "abc{\n}"
-        };
-        assert_eq!(format!("{s:?}"), expected);
-
-        let obj: Id<Object, Owned> = unsafe { Id::cast(NSObject::new()) };
-        let ptr: *const Object = &*obj;
-        let s = unsafe {
-            NSAttributedString::new_with_attributes(
-                &NSString::from_str("abc"),
-                &NSDictionary::from_keys_and_objects(&[&*NSString::from_str("test")], vec![obj]),
-            )
-        };
-        let expected = if cfg!(feature = "gnustep-1-7") {
-            format!("abc{{test = \"<NSObject: {ptr:?}>\"; }}")
-        } else {
-            format!("abc{{\n    test = \"<NSObject: {ptr:?}>\";\n}}")
-        };
-        assert_eq!(format!("{s:?}"), expected);
     }
 }
