+++ crates/icrate/src/Foundation/additions/bundle.rs
@@ -1,9 +1,8 @@
+#![cfg(feature = "Foundation_NSBundle")]
 use core::panic::{RefUnwindSafe, UnwindSafe};
 
-use objc2::rc::{Id, Shared};
-use objc2::runtime::Object;
-
-use crate::Foundation::{NSBundle, NSCopying, NSString};
+use crate::common::*;
+use crate::Foundation::{self, NSBundle};
 
 // SAFETY: Bundles are documented as thread-safe.
 unsafe impl Sync for NSBundle {}
@@ -13,7 +12,11 @@ impl UnwindSafe for NSBundle {}
 impl RefUnwindSafe for NSBundle {}
 
 impl NSBundle {
-    pub fn name(&self) -> Option<Id<NSString, Shared>> {
+    #[cfg(feature = "Foundation_NSString")]
+    #[cfg(feature = "Foundation_NSDictionary")]
+    pub fn name(&self) -> Option<Id<Foundation::NSString, Shared>> {
+        use Foundation::{NSCopying, NSString};
+
         let info = self.infoDictionary()?;
         // TODO: Use ns_string!
         let name = info.get(&NSString::from_str("CFBundleName"))?;
@@ -22,23 +25,5 @@ impl NSBundle {
         // SAFETY: TODO
         let name = unsafe { ptr.as_ref().unwrap_unchecked() };
         Some(name.copy())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use alloc::format;
-    use std::println;
-
-    #[test]
-    #[cfg_attr(not(target_os = "macos"), ignore = "varies between platforms")]
-    fn try_running_functions() {
-        // This is mostly empty since cargo doesn't bundle the application
-        // before executing.
-        let bundle = NSBundle::mainBundle();
-        println!("{bundle:?}");
-        assert_eq!(format!("{:?}", bundle.infoDictionary().unwrap()), "{}");
-        assert_eq!(bundle.name(), None);
     }
 }
