+++ crates/icrate/src/Foundation/additions/lock.rs
@@ -1,6 +1,7 @@
-use objc2::rc::{Id, Owned};
-use objc2::{extern_methods, ConformsTo};
+#![cfg(feature = "Foundation_NSLock")]
+use objc2::ConformsTo;
 
+use crate::common::*;
 use crate::Foundation::{NSLock, NSLocking};
 
 // TODO: Proper Send/Sync impls here
@@ -13,21 +14,3 @@ extern_methods!(
         pub fn new() -> Id<Self, Owned>;
     }
 );
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn lock_unlock() {
-        let lock = NSLock::new();
-        let locking: &NSLocking = lock.as_protocol();
-        unsafe {
-            locking.lock();
-            assert!(!lock.tryLock());
-            locking.unlock();
-            assert!(lock.tryLock());
-            locking.unlock();
-        }
-    }
-}
