+++ crates/icrate/src/Foundation/additions/process_info.rs
@@ -1,3 +1,4 @@
+#![cfg(feature = "Foundation_NSProcessInfo")]
 use core::fmt;
 use core::panic::{RefUnwindSafe, UnwindSafe};
 
@@ -11,27 +12,11 @@ unsafe impl Sync for NSProcessInfo {}
 impl UnwindSafe for NSProcessInfo {}
 impl RefUnwindSafe for NSProcessInfo {}
 
+#[cfg(feature = "Foundation_NSString")]
 impl fmt::Debug for NSProcessInfo {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.debug_struct("NSProcessInfo")
             .field("processName", &self.processName())
             .finish_non_exhaustive()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use alloc::format;
-
-    use super::*;
-
-    #[test]
-    fn test_debug() {
-        let info = NSProcessInfo::processInfo();
-        let expected = format!(
-            "NSProcessInfo {{ processName: {:?}, .. }}",
-            info.processName()
-        );
-        assert_eq!(format!("{info:?}"), expected);
     }
 }
